requirements.txt
# Core
aiohttp>=3.9
asyncio-throttle>=1.0

# Email parsing
mail-parser>=3.15
dkimpy>=1.1
checkdmarc>=5.0
dnspython>=2.4

# QR / Image
pyzbar>=0.1.9
opencv-python-headless>=4.8
Pillow>=10.0
PyMuPDF>=1.23
python-docx>=1.0

# URL / Domain
python-whois>=0.9
tldextract>=5.0

# Browser automation (URL detonation)
playwright>=1.40

# NLP / ML
scikit-learn>=1.3
joblib>=1.3
numpy>=1.25

# API framework
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.5

# Storage
sqlalchemy>=2.0
aiosqlite>=0.19

# IOC export
stix2>=3.0

# Reporting
jinja2>=3.1

# Testing
pytest>=7.4
pytest-asyncio>=0.23
pytest-mock>=3.12
aioresponses>=0.7
httpx>=0.25

# Code quality
flake8>=6.1
mypy>=1.7

# QR code generation (for tests)
qrcode>=7.4
